Change360Network/requirements.txt
asgiref==3.2.10
Django==3.1.2
django-countries==6.1.3
django-filter==2.4.0
djangorestframework==3.12.1
flake8==3.8.4
importlib-metadata==2.0.0
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
sqlparse==0.4.1
zipp==3.3.0
